Shared 페이지로 분리 (2)
1. App -> SharedPage로 옮겼던 state들 다시 App으로 옮김
2. Nav랑 Footer 부분을 SharedPage로 옮김
