+++ part2/src/components/App.js
@@ -1,10 +1,12 @@
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { useEffect, useState } from "react";
-import { getSampleUser } from "../api";
-import Nav from "./Nav/Nav";
-import Footer from "./Footer/Footer";
+import { getSampleUser, getFolderInfo } from "../api";
+import SharedPage from "../pages/SharedPage";
+import FolderPage from "../pages/FolderPage";
 
-function App({ children }) {
+function App() {
   const [userInfo, setUserInfo] = useState({});
+  const [folderInfo, setFolderInfo] = useState({});
 
   const handleLoadUser = async () => {
     const result = await getSampleUser();
@@ -14,16 +16,31 @@ function App({ children }) {
     setUserInfo((prevInfo) => ({ ...prevInfo, email, profileImageSource }));
   };
 
+  const handleLoadFolder = async () => {
+    const result = await getFolderInfo();
+    if (!result) return;
+
+    const { folder } = result;
+    setFolderInfo((prevInfo) => ({
+      ...prevInfo,
+      owner: folder.owner,
+      name: folder.name,
+      links: folder.links,
+    }));
+  };
+
   useEffect(() => {
     handleLoadUser();
+    handleLoadFolder();
   }, []);
 
   return (
-    <>
-      <Nav email={userInfo.email} imgUrl={userInfo.profileImageSource} />
-      <div>{children}</div>
-      <Footer />
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/shared" element={<SharedPage userInfo={userInfo} folderInfo={folderInfo} />} />
+        <Route path="/folder" element={<FolderPage />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
