--- part2/src/components/MainContent/MainContent.js
@@ -1,16 +1,0 @@
-import CardList from "../CardList/CardList";
-import Search from "../Search/Search";
-import "./MainContent.css";
-
-function MainContent({ links }) {
-  return (
-    <div className="mainContent-wrapper">
-      <div className="mainContent">
-        <Search />
-        <CardList links={links} />
-      </div>
-    </div>
-  );
-}
-
-export default MainContent;
