+++ part2/src/pages/FolderPage.js
@@ -1,5 +1,9 @@
+import { useEffect, useState } from "react";
+import { getFolderInfo } from "../api";
+import Folder from "../components/Folder/Folder";
+
 function FolderPage() {
-  return <></>;
+  return <>{/* <MainContent links={folderInfo.links} /> */}</>;
 }
 
 export default FolderPage;
